Move `content.js` to component-specific folder
