+++ packages/components/src/accessible-popover/accessible-popover-content/README.md
@@ -1,0 +1,15 @@
+# AccessiblePopoverContent
+
+<div class="callout callout-alert">
+This feature is still experimental. “Experimental” means this is an early implementation subject to drastic and breaking changes.
+</div>
+
+`AccessiblePopoverContent` is a component used by `AccessiblePopover` to render its contents.
+
+## Usage
+
+TODO
+
+## Props
+
+TODO
