+++ packages/components/src/accessible-popover/accessible-popover-content/index.js
@@ -1,0 +1,1 @@
+export { default } from './component';
