+++ packages/components/src/accessible-popover/index.js
@@ -1,2 +1,4 @@
 export { default as AccessiblePopover } from './accessible-popover';
+// TODO: should the Content component be exported?
+export { default as AccessiblePopoverContent } from './accessible-popover-content';
 export * from './context';
